refactor(session): extract the attach controller into src/session/attach/
attachToProcess and detachFromProcess move verbatim onto AttachController,
wired as AttachController(ctx, proxyLauncher, breakpoints) over a new
AttachContext slice of OperationsContext (tunables read late-bound, so the
tests that shrink attachVerifyTimeoutMs on the facade still take effect). The
facade keeps one-line delegates for both.

The two inline blocks inside attachToProcess become named functions in
attach/attach-verification.ts with typed inputs: verifyAttachThreads (the
'threads' poll with the proxy-gone latch, returning {ok, threads} |
{ok:false, proxyGone, lastFailure}; the reason/teardown/throw stays in the
controller) and pauseAfterAttach (the post-attach pause bounded by the
attachPauseStopTimeoutMs tunable). sendThreadsRequestBounded moves alongside
them as a free function. Every log line is unchanged.
