+++ src/session/session-manager-operations.ts
@@ -7,13 +7,8 @@
 import {
   Breakpoint,
   FunctionBreakpoint,
-  SessionState,
-  SessionLifecycleState,
   type ExceptionBreakMode
 } from '@debugmcp/shared';
-import { ManagedSession } from './session-store.js';
-import { DebugProtocol } from '@vscode/debugprotocol';
-import { ErrorMessages } from '../utils/error-messages.js';
 import { SessionManagerData } from './session-manager-data.js';
 import type { OperationsContext } from './operations-context.js';
 import { BreakpointController } from './breakpoints/breakpoint-controller.js';
@@ -33,7 +28,7 @@ import {
 } from './mirror/mirror-controller.js';
 import { ProxyLauncher } from './launch/proxy-launcher.js';
 import { DebugLauncher } from './launch/debug-launcher.js';
-import { logProxyFailure } from './launch/proxy-failure-diagnostics.js';
+import { AttachController } from './attach/attach-controller.js';
 import { CustomLaunchRequestArguments, DebugResult } from './session-manager-core.js';
 
 /** Result type for evaluate expression operations. */
@@ -150,6 +145,7 @@ export abstract class SessionManagerOperations extends SessionManagerData {
   protected readonly mirror = new MirrorController(this.opsContext);
   protected readonly proxyLauncher = new ProxyLauncher(this.opsContext);
   protected readonly launcher = new DebugLauncher(this.opsContext, this.proxyLauncher, this.breakpoints);
+  protected readonly attach = new AttachController(this.opsContext, this.proxyLauncher, this.breakpoints);
 
   /**
    * Start (or dry-run) a launch-mode debug session. Delegates to the launcher,
@@ -295,7 +291,8 @@ export abstract class SessionManagerOperations extends SessionManagerData {
   }
 
   /**
-   * Attach to a running process for debugging
+   * Attach to a running process. Delegates to the attach controller, which
+   * owns the attach sequence including thread verification.
    */
   async attachToProcess(
     sessionId: string,
@@ -312,513 +309,17 @@ export abstract class SessionManagerOperations extends SessionManagerData {
       adapterConfig?: Record<string, unknown>;
     }
   ): Promise<DebugResult> {
-    const session = this._getSessionById(sessionId);
-    this.logger.info(
-      `[SessionManager] Attempting to attach to process for session ${sessionId}`,
-      attachConfig
-    );
-
-    // The verification-window override is consumed by the thread-discovery
-    // loop below, not by the adapter — strip it from the config that becomes
-    // the DAP attach arguments. Validate before any state mutation.
-    // breakOnExceptions maps to setExceptionBreakpoints, not attach args —
-    // strip it too and thread it through the proxy config instead.
-    // adapterConfig is merged by startProxyManager (the same slot launch uses
-    // for adapterLaunchConfig, issue #336) — strip it here so the wrapper key
-    // itself cannot leak into the DAP attach arguments.
-    const { verifyTimeout, breakOnExceptions, adapterConfig, ...adapterAttachConfig } = attachConfig;
-    if (adapterConfig && adapterConfig.stopOnEntry !== undefined) {
-      this.logger.warn(
-        '[SessionManager] adapterConfig.stopOnEntry reaches the adapter but does not affect post-attach pause verification; prefer the top-level stopOnEntry parameter'
-      );
-    }
-    let verifyTimeoutOverride = verifyTimeout;
-    if (verifyTimeoutOverride !== undefined) {
-      if (
-        typeof verifyTimeoutOverride !== 'number' ||
-        !Number.isFinite(verifyTimeoutOverride) ||
-        verifyTimeoutOverride <= 0
-      ) {
-        return {
-          success: false,
-          state: session.state,
-          error: `'verifyTimeout' must be a positive number of milliseconds, got: ${String(verifyTimeoutOverride)}`
-        };
-      }
-      const maxVerifyTimeoutMs = 600000;
-      if (verifyTimeoutOverride > maxVerifyTimeoutMs) {
-        this.logger.warn(
-          `[SessionManager] verifyTimeout ${verifyTimeoutOverride}ms exceeds the maximum; clamping to ${maxVerifyTimeoutMs}ms`
-        );
-        verifyTimeoutOverride = maxVerifyTimeoutMs;
-      }
-    }
-
-    // Languages whose adapter declares no attach implementation fail fast,
-    // before any state mutation (issue #331). Only an explicit 'none'
-    // declaration is enforced — absent metadata falls through to the
-    // adapter's natural behavior.
-    // getFactoryMetadata is on IAdapterRegistry (issue #435 part 4); the
-    // runtime guard stays for partial registry doubles, but skipping the
-    // gate must never be silent — that is the fail-open degradation the
-    // typed surface exists to expose.
-    if (typeof this.adapterRegistry.getFactoryMetadata === 'function') {
-      const factoryMeta = await this.adapterRegistry.getFactoryMetadata(session.language).catch(() => undefined);
-      if (factoryMeta?.modes?.attach === 'none') {
-        return {
-          success: false,
-          state: session.state,
-          error: ErrorMessages.attachModeNotSupported(session.language)
-        };
-      }
-    } else {
-      this.logger.warn(
-        `[SessionManager] adapterRegistry has no getFactoryMetadata; skipping the attach-'none' ` +
-          `enforcement gate for '${session.language}'.`
-      );
-    }
-
-    if (session.proxyManager) {
-      this.logger.warn(
-        `[SessionManager] Session ${sessionId} already has an active proxy. Terminating before attaching.`
-      );
-      // Session-preserving teardown (same landmine as startDebugging, #238)
-      await this.stopProxyPreservingSession(session);
-    }
-
-    // Update to INITIALIZING state and set lifecycle to ACTIVE
-    this._updateSessionState(session, SessionState.INITIALIZING);
-    this.sessionStore.update(sessionId, {
-      sessionLifecycle: SessionLifecycleState.ACTIVE,
-      attachMode: true,
-    });
-
-    try {
-      // For attach mode, we use a placeholder scriptPath
-      // The actual attach logic will be handled by the adapter via dapLaunchArgs
-      const placeholderPath = 'attach://remote';
-
-      // Pass attach config through dapLaunchArgs with special request type
-      const attachLaunchArgs = {
-        ...adapterAttachConfig,
-        request: 'attach',
-        __attachMode: true  // Internal flag to signal attach mode
-      };
-
-      // Attach never receives a policy default (issue #244) — record the
-      // user's value (possibly undefined) for read-back symmetry.
-      session.effectiveBreakOnExceptions = breakOnExceptions;
-
-      const attachConfigData = await this.proxyLauncher.start(
-        session,
-        placeholderPath,
-        undefined,
-        attachLaunchArgs as Partial<CustomLaunchRequestArguments>,
-        false,
-        adapterConfig,  // merged over the attach config before transformAttachConfig (issue #336)
-        breakOnExceptions
-      );
-
-      // Perform language-specific handshake if required, mirroring
-      // startDebugging. For js-debug the whole DAP sequence — initialize,
-      // configurationDone and the DAP 'attach' request itself — is driven
-      // here because the proxy worker skips its built-in attach flow for
-      // command-queueing policies. Policies without performHandshake are
-      // untouched: their attach is performed by the proxy worker.
-      const policy = this.selectPolicy(session.language);
-      if (policy.performHandshake) {
-        try {
-          await policy.performHandshake({
-            proxyManager: session.proxyManager,
-            sessionId: session.id,
-            dapLaunchArgs: attachLaunchArgs as Partial<CustomLaunchRequestArguments>,
-            scriptPath: placeholderPath,
-            scriptArgs: undefined,
-            breakpoints: session.breakpoints,
-            launchConfig: attachConfigData,
-            breakOnExceptions
-          });
-        } catch (handshakeErr) {
-          this.logger.warn(
-            `[SessionManager] Language handshake for attach returned with warning/error: ${
-              handshakeErr instanceof Error ? handshakeErr.message : String(handshakeErr)
-            }`
-          );
-        }
-      }
-
-      // Set session state based on stopOnEntry
-      let finalState = session.state;
-
-      if (attachConfig.stopOnEntry !== false) {
-        // Verify the attach actually produced a debuggable target before
-        // reporting PAUSED: poll DAP 'threads' until the debugger reports at
-        // least one thread. A debugger that cannot enumerate any threads after
-        // attach is not usable — reporting success would be a lie (issue #124:
-        // JS attach reported success + "paused" while the js-debug child
-        // session never connected to the target).
-        if (!session.proxyManager) {
-          throw new Error('Proxy manager is not available after attach initialization');
-        }
-        const proxyManager = session.proxyManager;
-
-        const verifyTimeoutMs = verifyTimeoutOverride ?? this.attachVerifyTimeoutMs;
-        const pollIntervalMs = this.attachVerifyIntervalMs;
-        const deadline = Date.now() + verifyTimeoutMs;
-
-        let threads: DebugProtocol.Thread[] | undefined;
-        let lastFailure = 'no threads response received';
-
-        // Some adapters reject the attach only after reporting themselves
-        // configured (CodeLLDB does this for e.g. ptrace EPERM). The proxy
-        // then dies mid-verify and every remaining 'threads' poll would throw
-        // a generic "Proxy not initialized", masking the real error — so latch
-        // the first proxy error/exit as the failure and stop polling.
-        let proxyGone = false;
-        const onProxyError = (err: Error): void => {
-          if (!proxyGone) {
-            proxyGone = true;
-            lastFailure = err.message;
-          }
-        };
-        const onProxyExit = (code: number | null, signal?: string): void => {
-          if (!proxyGone) {
-            proxyGone = true;
-            lastFailure = `debug adapter exited during attach verification (code=${code}${signal ? `, signal=${signal}` : ''})`;
-          }
-        };
-        proxyManager.on('error', onProxyError);
-        proxyManager.on('exit', onProxyExit);
-
-        const requestThreads = async (): Promise<void> => {
-          const remainingMs = Math.max(deadline - Date.now(), 1);
-          const threadsResponse = await this.sendThreadsRequestBounded(proxyManager, remainingMs);
-          if (proxyGone) {
-            return;
-          }
-          if (threadsResponse?.success === false) {
-            lastFailure = threadsResponse.message || `'threads' request failed`;
-            return;
-          }
-          const reported = threadsResponse?.body?.threads;
-          if (Array.isArray(reported) && reported.length > 0) {
-            threads = reported;
-          } else {
-            lastFailure = 'debugger reported zero threads';
-          }
-        };
-
-        try {
-          // First discovery attempt.
-          try {
-            await requestThreads();
-          } catch (err) {
-            if (!proxyGone) {
-              lastFailure = err instanceof Error ? err.message : String(err);
-            }
-            this.logger.warn(`[SessionManager] Initial thread discovery for attach failed: ${lastFailure}`);
-          }
-
-          // Retry until the deadline if the debugger has not reported threads yet.
-          while (!threads && !proxyGone && Date.now() < deadline) {
-            const sleepMs = Math.min(pollIntervalMs, Math.max(deadline - Date.now(), 1));
-            await new Promise((resolve) => setTimeout(resolve, sleepMs));
-            if (proxyGone || Date.now() >= deadline) {
-              break;
-            }
-            try {
-              await requestThreads();
-            } catch (err) {
-              if (!proxyGone) {
-                lastFailure = err instanceof Error ? err.message : String(err);
-              }
-            }
-          }
-        } finally {
-          proxyManager.removeListener('error', onProxyError);
-          proxyManager.removeListener('exit', onProxyExit);
-        }
-
-        if (!threads) {
-          const reason = proxyGone
-            ? ErrorMessages.attachAdapterFailed(lastFailure)
-            : ErrorMessages.attachVerifyFailed(verifyTimeoutMs, lastFailure);
-          this.logger.error(`[SessionManager] ${reason} — tearing down proxy for session ${sessionId}`);
-          // Tear down the proxy using the same mechanics as closeSession, but
-          // keep the session record so the failure is inspectable as ERROR.
-          try {
-            this.cleanupProxyEventHandlers(session, proxyManager);
-          } catch (cleanupError) {
-            this.logger.error(`[SessionManager] Error during listener cleanup for failed attach:`, cleanupError);
-          }
-          try {
-            await proxyManager.stop();
-          } catch (stopError) {
-            this.logger.error(`[SessionManager] Error stopping proxy for failed attach:`, stopError);
-          } finally {
-            session.proxyManager = undefined;
-          }
-          throw new Error(reason);
-        }
-
-        // Prefer a thread named "main" (common in JVM debugging)
-        const mainThread = threads.find(t => t.name === 'main');
-        const discoveredThreadId = mainThread ? mainThread.id : threads[0].id;
-        this.logger.info(`[SessionManager] Discovered ${threads.length} threads. Using threadId=${discoveredThreadId} (name=${mainThread?.name || threads[0].name})`);
-        proxyManager.setCurrentThreadId(discoveredThreadId);
-        this.logger.info(`[SessionManager] Set threadId=${discoveredThreadId} for attach mode`);
-
-        // Some debuggers (rdbg; js-debug attaches with continueOnAttach) do
-        // not suspend a running target on attach; issue an explicit pause so
-        // the PAUSED state we report is real, and wait for the stop to be
-        // observed before reporting it. Sent after thread verification so it
-        // reaches the debuggee-owning session (for js-debug the pause is
-        // routed to the child session, which exists once threads are
-        // reported). A rejected pause means the target is already stopped
-        // (e.g. started suspended) — fine, no stop event will follow.
-        const attachBehavior = this.selectPolicy(session.language).getAttachBehavior?.();
-        if (attachBehavior?.pauseAfterAttach) {
-          let stopSettled = false;
-          let stopTimer: ReturnType<typeof setTimeout> | undefined;
-          let onStopped: (() => void) | undefined;
-          const stoppedSeen = new Promise<boolean>((resolve) => {
-            onStopped = () => {
-              if (!stopSettled) {
-                stopSettled = true;
-                if (stopTimer) clearTimeout(stopTimer);
-                resolve(true);
-              }
-            };
-            stopTimer = setTimeout(() => {
-              if (!stopSettled) {
-                stopSettled = true;
-                resolve(false);
-              }
-            }, this.attachPauseStopTimeoutMs);
-            proxyManager.once('stopped', onStopped);
-          });
-          try {
-            // pauseAllThreads (issue #465): threadId 0 asks the adapter for a
-            // process-wide suspend — the JDI bridge then re-anchors its
-            // stopped event to a thread that can actually report frames,
-            // instead of single-thread-suspending whichever id we picked.
-            const pauseThreadId = attachBehavior.pauseAllThreads ? 0 : discoveredThreadId;
-            await proxyManager.sendDapRequest('pause', { threadId: pauseThreadId });
-            this.logger.info(`[SessionManager] Sent post-attach pause (threadId=${pauseThreadId})`);
-            const stopObserved = await stoppedSeen;
-            if (!stopObserved) {
-              this.logger.warn(
-                `[SessionManager] No 'stopped' event within ${this.attachPauseStopTimeoutMs}ms after post-attach pause; reported state may lag the engine`
-              );
-            }
-          } catch (err) {
-            // Already stopped (e.g. target was started suspended) — fine.
-            this.logger.info(
-              `[SessionManager] Post-attach pause not needed/accepted: ${err instanceof Error ? err.message : String(err)}`
-            );
-          } finally {
-            stopSettled = true;
-            if (stopTimer) clearTimeout(stopTimer);
-            if (onStopped) proxyManager.removeListener('stopped', onStopped);
-          }
-        }
-
-        this._updateSessionState(session, SessionState.PAUSED);
-        finalState = SessionState.PAUSED;
-        this.logger.info(`[SessionManager] Set session ${sessionId} to PAUSED after attach (stopOnEntry=${attachConfig.stopOnEntry})`);
-      } else {
-        // JVM is already running (suspend=n), set RUNNING state
-        this._updateSessionState(session, SessionState.RUNNING);
-        finalState = SessionState.RUNNING;
-        this.logger.info(`[SessionManager] Set session ${sessionId} to RUNNING (stopOnEntry=false, process started with suspend=n)`);
-      }
-
-      // Attach parity with the post-launch belt-and-braces re-sync (issues
-      // #236/#439, here for #500): breakpoints set before attach_to_process
-      // were sent during the handshake, whose responses the policy may
-      // discard — and for js-debug the child session only answered its
-      // pending stub while adoption was in flight. Re-sending now, with the
-      // debuggee-owning session provably live, delivers the authoritative
-      // verification. Replace-all with the identical set is idempotent;
-      // syncBreakpointsForFile never throws and no-ops unless live.
-      if (session.breakpoints.size > 0) {
-        const files = [...new Set(Array.from(session.breakpoints.values()).map(bp => bp.file))];
-        for (const file of files) {
-          // forceFreshEcho: js-debug answers a no-change re-send with an
-          // empty echo, and pre-attach breakpoints were already registered
-          // via its pending-target queue — without a fresh echo their
-          // verified state is unrecoverable (issue #500).
-          await this.breakpoints.syncBreakpointsForFile(session, file, { forceFreshEcho: true });
-        }
-      }
-      if ((session.functionBreakpoints?.size ?? 0) > 0) {
-        await this.breakpoints.syncFunctionBreakpoints(session);
-      }
-      // Unverified-at-attach function breakpoints get the same launch-style
-      // warning (issue #308); bind-late adapters (js/java) stay suppressed
-      // inside the builder.
-      const attachFnBpWarning = this.breakpoints.functionBreakpointLaunchWarning(session);
-
-      const attachData: Record<string, unknown> = {
-        message: attachConfig.processId
-          ? `Attached to process PID ${attachConfig.processId}`
-          : `Attached to process at ${attachConfig.host || 'localhost'}:${attachConfig.port}`,
-        attachConfig
-      };
-      // Surface adapterConfig keys the adapter's attach transform dropped
-      // (issue #450) and keys forwarded to the adapter unrecognized (issue
-      // #466) — "unknown attach keys should either work or warn".
-      const droppedKeys = session.attachDroppedConfigKeys;
-      const forwardedKeys = session.attachForwardedUnknownConfigKeys;
-      session.attachDroppedConfigKeys = undefined;
-      session.attachForwardedUnknownConfigKeys = undefined;
-      const warningParts: string[] = [];
-      if (attachFnBpWarning) {
-        warningParts.push(attachFnBpWarning);
-      }
-      if (droppedKeys && droppedKeys.length > 0) {
-        warningParts.push(
-          `adapterConfig key(s) not supported by the ${session.language} attach request were ignored: ${droppedKeys.join(', ')}`
-        );
-      }
-      if (forwardedKeys && forwardedKeys.length > 0) {
-        warningParts.push(
-          `adapterConfig key(s) not recognized by mcp-debugger were forwarded to the ${session.language} adapter as-is: ${forwardedKeys.join(', ')}`
-        );
-      }
-      if (warningParts.length > 0) {
-        attachData.warning = warningParts.join('; ');
-      }
-
-      return {
-        success: true,
-        state: finalState,
-        data: attachData
-      };
-    } catch (error) {
-      this.logger.error(`[SessionManager] Failed to attach to process for session ${sessionId}:`, error);
-      // Never leave a live proxy chain behind a failed attach — e.g.
-      // ProxyManager.start()'s init timeout rejects after the worker was
-      // spawned (issue #337). Idempotent with the verify-failure teardown
-      // above, which already nulled session.proxyManager.
-      await this.stopProxyPreservingSession(session);
-      this._updateSessionState(session, SessionState.ERROR);
-
-      // Surface the same structured diagnostics the launch path returns
-      // (issue #551) and log the same full failure record it logs, proxy-log
-      // tail included (issue #561) — an attach that dies during proxy
-      // initialization used to leave the adapter's own complaint unreadable.
-      // Teardown only clears the proxy handle; logDir and the error's
-      // initProgress survive it, so this reads after the teardown and can
-      // never keep it from running.
-      const diagnosticData = await logProxyFailure(
-        { logger: this.logger, fileSystem: this.fileSystem },
-        session,
-        error,
-        'attachToProcess'
-      );
-      const message = error instanceof Error ? error.message : String(error);
-      return {
-        success: false,
-        state: SessionState.ERROR,
-        error: `Failed to attach: ${message}`,
-        ...(Object.keys(diagnosticData).length > 0 ? { data: diagnosticData } : {})
-      };
-    }
+    return this.attach.attachToProcess(sessionId, attachConfig);
   }
 
   /**
-   * Send a DAP 'threads' request bounded by a timeout so a hung request
-   * cannot stall attach verification past its deadline. The underlying
-   * request keeps its own lifecycle; only the wait here is bounded.
-   */
-  private async sendThreadsRequestBounded(
-    proxyManager: NonNullable<ManagedSession['proxyManager']>,
-    timeoutMs: number
-  ): Promise<DebugProtocol.ThreadsResponse | undefined> {
-    let timer: ReturnType<typeof setTimeout> | undefined;
-    try {
-      return await Promise.race([
-        proxyManager.sendDapRequest<DebugProtocol.ThreadsResponse>('threads', {}),
-        new Promise<never>((_, reject) => {
-          timer = setTimeout(
-            () => reject(new Error(`'threads' request did not respond within ${timeoutMs}ms`)),
-            timeoutMs
-          );
-        })
-      ]);
-    } finally {
-      if (timer) {
-        clearTimeout(timer);
-      }
-    }
-  }
-
-  /**
-   * Detach from the debugged process without terminating it
+   * Detach from the debugged process without terminating it.
    */
   async detachFromProcess(
     sessionId: string,
     terminateProcess: boolean = false
   ): Promise<DebugResult> {
-    const session = this._getSessionById(sessionId);
-    this.logger.info(
-      `[SessionManager] Detaching from process for session ${sessionId}, terminateProcess: ${terminateProcess}`
-    );
-
-    if (!session.proxyManager) {
-      return {
-        success: false,
-        state: session.state,
-        error: 'No active debug session to detach from'
-      };
-    }
-
-    try {
-      if (terminateProcess) {
-        // Terminate the process
-        await this.closeSession(sessionId);
-      } else {
-        // Disconnect without terminating - send DAP disconnect request
-        try {
-          await session.proxyManager.sendDapRequest('disconnect', {
-            terminateDebuggee: false
-          });
-        } catch (disconnectError) {
-          this.logger.warn(`[SessionManager] Disconnect request failed, continuing with cleanup:`, disconnectError);
-        }
-
-        // Stop the proxy manager — it may already be gone if the disconnect
-        // request triggered a 'terminated' event that cleared proxyManager.
-        if (session.proxyManager) {
-          await session.proxyManager.stop();
-        }
-
-        this._updateSessionState(session, SessionState.STOPPED);
-        this.sessionStore.update(sessionId, {
-          sessionLifecycle: SessionLifecycleState.TERMINATED
-        });
-      }
-
-      return {
-        success: true,
-        state: SessionState.STOPPED,
-        data: {
-          message: terminateProcess
-            ? 'Detached and terminated process'
-            : 'Detached from process (process still running)'
-        }
-      };
-    } catch (error) {
-      this.logger.error(`[SessionManager] Failed to detach from process for session ${sessionId}:`, error);
-
-      const message = error instanceof Error ? error.message : String(error);
-      return {
-        success: false,
-        state: session.state,
-        error: `Failed to detach: ${message}`
-      };
-    }
+    return this.attach.detachFromProcess(sessionId, terminateProcess);
   }
 
   /**
